Rewrite all tests using the Catch2 testing framework

+++ delta/modulo_w.h
@@ -13,7 +13,7 @@ vector<vector<pair<int, int>>> classify_coords(int w) {
 }
 
 int delta_modulo_w(Factors f, int w, vector<vector<pair<int,int>>> classes) {
-    int total_visited = 0;
+    uint64_t total_visited = 0;
 
     // Determine coordinates based on the modulus and then
     // determine the lower bounds of each vector
@@ -23,8 +23,8 @@ int delta_modulo_w(Factors f, int w, vector<vector<pair<int,int>>> classes) {
     // m is the modulus
     for (int m = 0; m < classes.size(); m++) {
 
-        int visited_range = f.product_bound() - lower_bounds[m];
-        int visited_size = visited_range / w + 1;
+        uint64_t visited_range = f.product_bound() - lower_bounds[m];
+        uint64_t visited_size = visited_range / w + 1;
         boost::dynamic_bitset<> visited(visited_size);
 
         // Free stuff!
@@ -37,9 +37,9 @@ int delta_modulo_w(Factors f, int w, vector<vector<pair<int,int>>> classes) {
             pair<int,int> coords = classes[m][p];
 
             // The starting row.  We could skip the first few here.
-            int row_start = coords.second == 0 ? w : coords.second;
+            uint64_t row_start = coords.second == 0 ? w : coords.second;
 
-            for (int r = row_start, t = 1; r < f.row_bound(); r += w) {
+            for (uint64_t r = row_start, t = 1; r < f.row_bound(); r += w) {
                 
                 while (r >= f.row_bound_on_pair(t)) {
                     t++;
@@ -48,20 +48,22 @@ int delta_modulo_w(Factors f, int w, vector<vector<pair<int,int>>> classes) {
                 // Start on the coordinate point, and increment by w
                 // until passed the row number (we only scan the lower half)
                 // and passed the lower bound, we get those for free already
-                int start = coords.first;
+                uint64_t start = coords.first;
                 while (start < r || start <= lower_bounds[m] / r) start += w;
 
                 // The slow intuitive way:
-                /*
-                    for (int c = start; c < f.col_bound_on_pair(t); c += w) {
+                    for (uint64_t c = start; c < f.col_bound_on_pair(t); c += w) {
                         int p = c * r;                      // product
                         int i = (p - lower_bounds[m]) / w;  // index after translation and scaling
                         visited[i] = true;
                     }
+                /*
                     for (int p = start * r; p < f.col_bound_on_pair(t) * r; p += w * r) {
                         int i = (p - lower_bounds[m]) / w;  // index after translation and scaling
                         visited[i] = true;
                     }
+                */
+                /*
                     for (int p = start * r - lower_bounds[m]; 
                         p < f.col_bound_on_pair(t) * r - lower_bounds[m]; 
                         p += w * r
@@ -73,14 +75,16 @@ int delta_modulo_w(Factors f, int w, vector<vector<pair<int,int>>> classes) {
                 // The quicker way:
                 
                 // Need to figure out why this works
-                int offset = (start * r - lower_bounds[m]) % w;
+                /*
+                uint64_t offset = (start * r - lower_bounds[m]) % w;
 
-                for (int i = (start * r - lower_bounds[m]) / w; 
-                    i < ((int)f.col_bound_on_pair(t) * r - lower_bounds[m] - offset + w - 1) / w; 
+                for (uint64_t i = (start * r - lower_bounds[m]) / w; 
+                    i < ((uint64_t)f.col_bound_on_pair(t) * r - lower_bounds[m] - offset + w - 1) / w; 
                     i += r 
                 ) {
                     visited[i] = true;
                 }
+                */
             }
         }
